+++ src/components/DevtoolsPane/index.tsx
@@ -237,14 +237,16 @@ const RowSignalingOptions: React.FC = () => {
 
 const RowAdvancedSignalingOptions: React.FC = () => {
   const [collapsed, setCollapsed] = useState(true)
-  const audioStreamingLanguageCode = useAppSelector((state) => state.audioStreamingLanguageCode)
+  const enableAudioStreamingLanguageCode = useAppSelector(
+    (state) => state.enabledAudioStreamingLanguageCode,
+  )
   const audioLyraParamsBitrate = useAppSelector((state) => state.audioLyraParamsBitrate)
   const enabledVideoVP9Params = useAppSelector((state) => state.enabledVideoVP9Params)
   const enabledVideoH264Params = useAppSelector((state) => state.enabledVideoH264Params)
   const enabledVideoH265Params = useAppSelector((state) => state.enabledVideoH265Params)
   const enabledVideoAV1Params = useAppSelector((state) => state.enabledVideoAV1Params)
   const enabledOptions = [
-    audioStreamingLanguageCode !== '',
+    enableAudioStreamingLanguageCode,
     audioLyraParamsBitrate !== '',
     enabledVideoVP9Params,
     enabledVideoH264Params,